+++ src/main/java/com/imsweb/datagenerator/naaccr/rule/patient/PatientIdRule.java
@@ -23,7 +23,7 @@ public class PatientIdRule extends NaaccrDataGeneratorRule {
     }
 
     @Override
-    public void execute(Map<String, String> record, List<Map<String, String>> otherRecords, NaaccrDataGeneratorOptions options) {
+    public void execute(Map<String, String> record, List<Map<String, String>> otherRecords, NaaccrDataGeneratorOptions options, Map<String, Object> context) {
         record.put("patientIdNumber", String.format("%08d", _currentValue++));
     }
 }