+++ src/main/java/com/imsweb/datagenerator/naaccr/rule/tumor/DateOfDiagnosisRule.java
@@ -10,6 +10,9 @@ import com.imsweb.datagenerator.naaccr.NaaccrDataGeneratorOptions;
 import com.imsweb.datagenerator.naaccr.NaaccrDataGeneratorRule;
 import com.imsweb.datagenerator.utils.RandomUtils;
 
+import static com.imsweb.datagenerator.naaccr.NaaccrDataGenerator.CONTEXT_FLAG_AGE_GROUP_MAP;
+import static com.imsweb.datagenerator.naaccr.NaaccrDataGenerator.CONTEXT_FLAG_CURRENT_TUMOR_INDEX;
+
 public class DateOfDiagnosisRule extends NaaccrDataGeneratorRule {
 
     // unique identifier for this rule
@@ -23,7 +26,7 @@ public class DateOfDiagnosisRule extends NaaccrDataGeneratorRule {
     }
 
     @Override
-    public void execute(Map<String, String> record, List<Map<String, String>> otherRecords, NaaccrDataGeneratorOptions options) {
+    public void execute(Map<String, String> record, List<Map<String, String>> otherRecords, NaaccrDataGeneratorOptions options, Map<String, Object> context) {
 
         // latest possible date set only by options if defined
         Set<LocalDate> maxDxDates = new HashSet<>();
@@ -37,11 +40,33 @@ public class DateOfDiagnosisRule extends NaaccrDataGeneratorRule {
         if (propertyHasValue(record, "birthDateYear"))
             minDxDates.add(LocalDate.of(Integer.parseInt(record.get("birthDateYear")) + 1, 1, 1));
         // never go before dx date of patient's most recent tumor (if this isn't the first one)
-        if (!otherRecords.isEmpty()) {
+        if (otherRecords != null && !otherRecords.isEmpty()) {
             Map<String, String> lastTumor = otherRecords.get(otherRecords.size() - 1);
             minDxDates.add(LocalDate.of(Integer.parseInt(lastTumor.get("dateOfDiagnosisYear")), Integer.parseInt(lastTumor.get("dateOfDiagnosisMonth")),
                     Integer.parseInt(lastTumor.get("dateOfDiagnosisDay"))));
         }
+
+        if (context.get(CONTEXT_FLAG_CURRENT_TUMOR_INDEX) != null) {
+            Integer birthYear = Integer.parseInt(record.get("birthDateYear"));
+            Integer birthMonth = Integer.parseInt(record.get("birthDateMonth"));
+            Integer birthDay = Integer.parseInt(record.get("birthDateDay"));
+            LocalDate birthDate = LocalDate.of(birthYear, birthMonth, birthDay);
+
+            // PROBLEM: This brakes 3 previous rules:
+            // 1. Minimum date must be within 10 years of today.
+            // 2. Options specify a minimum DX date.
+            // 3. This tumor must be diagnosed after the previous ones for this patient.
+            // From Fabian: Only #2 is required. Try to get all tumors to use this minimum. If that can't be done, at least one tumor must meet it.
+            minDxDates.clear();
+            maxDxDates.clear();
+
+            int currentTumorIndex = (int)context.get(CONTEXT_FLAG_CURRENT_TUMOR_INDEX);
+            Map<Integer, Integer> ageGroupMap = (Map<Integer, Integer>)context.get(CONTEXT_FLAG_AGE_GROUP_MAP);
+            minDxDates.add(birthDate.plusYears((ageGroupMap.get(currentTumorIndex) * 10)));
+            maxDxDates.add(maxDate);
+        }
+
+
 
         LocalDate randomDate = RandomUtils.getRandomDateBetween(minDxDates, maxDxDates);