+++ src/main/java/com/imsweb/datagenerator/naaccr/rule/tumor/NapiiaRule.java
@@ -28,7 +28,7 @@ public class NapiiaRule extends NaaccrDataGeneratorRule {
     }
 
     @Override
-    public void execute(Map<String, String> record, List<Map<String, String>> otherRecords, NaaccrDataGeneratorOptions options) {
+    public void execute(Map<String, String> record, List<Map<String, String>> otherRecords, NaaccrDataGeneratorOptions options, Map<String, Object> context) {
         NapiiaInputRecordDto input = new NapiiaInputRecordDto();
         input.setRace1(record.get("race1"));
         input.setRace2(record.get("race2"));