+++ src/test/java/com/imsweb/datagenerator/naaccr/rule/patient/ComputedEthnicityRuleTest.java
@@ -13,12 +13,13 @@ public class ComputedEthnicityRuleTest {
     @Test
     public void testExecute() {
         Map<String, String> rec = new HashMap<>();
+        Map<String, Object> context = new HashMap<>();
 
         // computedEthnicity = 1 (non-hispanic last name, non-hispanic maiden name)
         rec.put("sex", "2");
         rec.put("nameLast", "JOHNSON");
         rec.put("nameMaiden", "MILLER");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("1"));
         rec = new HashMap<>();
 
@@ -26,7 +27,7 @@ public class ComputedEthnicityRuleTest {
         rec.put("sex", "1");
         rec.put("nameLast", "JOHNSON");
         rec.put("nameMaiden", "");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("2"));
         rec = new HashMap<>();
 
@@ -34,7 +35,7 @@ public class ComputedEthnicityRuleTest {
         rec.put("sex", "2");
         rec.put("nameLast", "JOHNSON");
         rec.put("nameMaiden", "");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("3"));
         rec = new HashMap<>();
 
@@ -42,7 +43,7 @@ public class ComputedEthnicityRuleTest {
         rec.put("sex", "2");
         rec.put("nameLast", "ZUZUARREGUI");
         rec.put("nameMaiden", "MILLER");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("4"));
         rec = new HashMap<>();
 
@@ -50,7 +51,7 @@ public class ComputedEthnicityRuleTest {
         rec.put("sex", "1");
         rec.put("nameLast", "ZUZUARREGUI");
         rec.put("nameMaiden", "");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("5"));
         rec = new HashMap<>();
 
@@ -58,7 +59,7 @@ public class ComputedEthnicityRuleTest {
         rec.put("sex", "2");
         rec.put("nameLast", "ZUZUARREGUI");
         rec.put("nameMaiden", "");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("6"));
         rec = new HashMap<>();
 
@@ -66,7 +67,7 @@ public class ComputedEthnicityRuleTest {
         rec.put("sex", "2");
         rec.put("nameLast", "");
         rec.put("nameMaiden", "ABAD");
-        _rule.execute(rec, null, null);
+        _rule.execute(rec, null, null, context);
         Assert.assertTrue(rec.get("computedEthnicity").equals("7"));
 
         // verify computed ethnicity source (constant, 2)