+++ src/test/java/com/imsweb/datagenerator/naaccr/rule/patient/SexRuleTest.java
@@ -21,7 +21,8 @@ public class SexRuleTest {
         // test the rule ten times, asserting that the execute() method always assigns a valid sex code to the patient
         for (int i = 0; i < 10; i++) {
             Map<String, String> rec = new HashMap<>();
-            _rule.execute(rec, null, null);
+            Map<String, Object> context = new HashMap<>();
+            _rule.execute(rec, null, null, context);
             Assert.assertTrue(validSex.contains(rec.get("sex")));
         }
     }